booking and logout fragment

+++ app/src/main/java/com/example/rentatease/activities/ApartmentListActivity.java
@@ -5,6 +5,8 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.text.TextUtils;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 
 import androidx.annotation.Nullable;
@@ -56,8 +58,8 @@ public class ApartmentListActivity extends AppCompatActivity {
                 for (DataSnapshot dataSnapshot : snapshot.getChildren()) {
 
                     Apartment apartment = dataSnapshot.getValue(Apartment.class);
-                    if(!apartment.isBooked())
-                    apartmentList.add(apartment);
+                    if (!apartment.isBooked())
+                        apartmentList.add(apartment);
                 }
 
                 adapter = new ApartmentRenteeAdapter(ApartmentListActivity.this, apartmentList);
@@ -72,21 +74,38 @@ public class ApartmentListActivity extends AppCompatActivity {
             }
         });
 
+
         btnLogout = findViewById(R.id.btnLogout);
         btnLogout.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                SharedPreferences sharedPreferences = getSharedPreferences(Const.SHAREDPREFERENCE, MODE_PRIVATE);
-                SharedPreferences.Editor editor = sharedPreferences.edit();
-                editor.clear();
-                editor.apply();
-                FirebaseAuth.getInstance().signOut();
-                Intent i = new Intent(ApartmentListActivity.this, LoginActivity.class);
-                startActivity(i);
-                finish();
+
             }
         });
 
+
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.navigation, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+        if (item.getItemId() == R.id.action_logout) {
+            SharedPreferences sharedPreferences = getSharedPreferences(Const.SHAREDPREFERENCE, MODE_PRIVATE);
+            SharedPreferences.Editor editor = sharedPreferences.edit();
+            editor.clear();
+            editor.apply();
+            FirebaseAuth.getInstance().signOut();
+            Intent i = new Intent(ApartmentListActivity.this, LoginActivity.class);
+            startActivity(i);
+            finish();
+        }
+        return true;
 
     }
 }
